Fix redefinition of Vector3f

+++ Air-Engine/src/Engine/Math/Vector3.hpp
@@ -260,6 +260,6 @@ namespace engine {
 		};
 
 		typedef Vector3<float> Vector3f;
-		typedef Vector3<double> Vector3f;
+		typedef Vector3<double> Vector3d;
 	}
 }
